requirements.txt
python-levenshtein
pandas
numpy